projet terminer sans les bonus

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: lide <lide@student.s19.be>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/03 17:25:13 by lide              #+#    #+#             */
-/*   Updated: 2022/01/24 17:33:30 by lide             ###   ########.fr       */
+/*   Updated: 2022/02/03 15:30:48 by lide             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
 	int	i;
